fix default namespace regression
